Cousas varias (teño que facer comits mais a miudo)
 * Creados mensaxes de error para as respostas de datos.
 * Creados metodos para actualizar mensaxes e canles na iu.
 * Creadas clases para mover datos de Canles e Mensaxes.
 * Creados metodos para formatear datos de mensaxes e canles.
 * Creado metodo actualizar que engloba o resto de metodos de actualizacion.
 * Modificados estados para que indique se o trafico esta aberto e/ou activo.
 * Modificado diagrama ER.
 * Borrado gitignore persoal.

+++ src/practicaacs/banco/bd/Conta.java
@@ -2,9 +2,9 @@ package practicaacs.banco.bd;
 
 public class Conta {
 	private int numero;
-	private float saldo;
+	private int saldo;
 	
-	public Conta(int numero, float saldo) {
+	public Conta(int numero, int saldo) {
 		super();
 		this.numero = numero;
 		this.saldo = saldo;
@@ -12,7 +12,7 @@ public class Conta {
 	public int getNumero() {
 		return numero;
 	}
-	public float getSaldo() {
+	public int getSaldo() {
 		return saldo;
 	}
 }
